requirements.txt
Django==4.0.6
djangorestframework==3.13.1
psycopg2==2.9.3
django-model-utils==4.2.0
